+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Add `buffer` function
 * Add `Signal#buffer` method
 * Add `apply` function
+* Add `Signal.throwError` static method
+* Add `catchError` function
+* Add `Signal#catchError` method
 
 ## 2.0.0 (2019-01-12)